[trusty] Add support for exposing handles to userspace
Adds a mechanism for exporting opaque integer handles to
userspace which can be used to represent various kernel
objects.

There is a per-context table that stores pointers to handle_t
objects. There's also a bitmap that is used for quick validation
of an integer handle from userspace as well as a fast search for
an empty handle id.

Change-Id: I1a6a87380db825287769213c58902437f7dba72b
Signed-off-by: Dima Zavin <dima@android.com>