[lib][syscall] A system call handling framework
libsyscall provides a table based framework for handling system calls. A
user of this library just provides a file named syscall_table.h with a
table of function pointers defined using the DEF_SYSCALL macro.

The system call vector handler provided as part of this library simply
jumps to the right function in the table based on system call number.

The system call ABI for a given architecture is documented in
arch/$(ARCH)/syscall.S

This library also provides a python script to generate system call stub
functions for userspace in GNU assembler syntax. The script can also be
used to sanity check a system call table file. For more info:

$ stubgen.py --help

Change-Id: I2c2c92af77333243e5723d705add3e7d1ee93312
Signed-off-by: Satya Popuri <popuri@google.com>