[trusty] Add an IPC mechanism for userspace trusty apps
This change introduces an IPC mechanism which may be used
by trusty userspace applications to interact with one another.

The IPC mechanism has two main constructs: ports and channels.
Ports are used by services to export a named path for a client to
connect to, e.g. "/service/foo" or
"/uuid/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx". When a client connects
to a port, it gets back a channel which it will use to interact
with the service. The service has to accept an incoming connection
request from the client, so it can get a chance to validate the
connection request if necessary. Once the connection has been
accepted, the server now gets a channel to communicate with the
client. In essense, ports are used to lookup services and then
the communication between server and client happens over a pair
of connected channels.

Once the communication has been established, the server and
client can communicate by exchanging messages. Presently,
only direct user buffers are supported which are copied to a
staging area in a receiving app. In the future, IPC clients
should be able to exchange buffers objects indirectly through
handles to be able to support large buffer passing.

For storage, we presently use a pre-allocated message area for
each receiver. The port creator specifies the size of the largest
message, and the number of outstanding messages that should be
supported on that port. When a connection is established,
those buffers are allocated for both channels in the
server/client pair.

Change-Id: I88a44f6187e7d4e165aa5e546b2ac034b971864e
Signed-off-by: Dima Zavin <dima@android.com>