[lib][uthread][arm] fix register allocation in user task startup
The code is using strdb/ldmdb instructions to setup the user mode
lr and pc. Previously, the code was allowing the compiler to
allocate two registers to that would be used to push lr/pc onto
the stack, and it would then restore them into user r13/r14. The
problem is that strdb/ldmdb do not retain the register list order
as the register list gets encoded as a bitmask in the instruction.
So, if compiler allocated, say, r7 for sp and r6 for lr, when restored
into user r13 and r14, the values would get reversed.

To fix, we force the values into specific registers (r2 and r3 at
the moment).

Change-Id: Ic843ab5bb2bcae63fcbb90763e403d0139af6b71
Signed-off-by: Dima Zavin <dima@android.com>