[lib][uthread] Replace strlcpy_from_user with strncpy_from_user
The existing implementation strlcpy_from_user has been silently
truncating string if provided buffer was not big enough to hold
it.

It is also somewhat dangerous to provide proper strlcpy-like
behavier for user space strings as it requires to return the
size of source string so we have to walk potentially unbound
data in user space. It is better to swith to strncpy like
behavier as it is sufficient for our needs as simpler to
implement.

Change-Id: I8dc07257087e4073f3e57772cb006a4a42f1c427