[lib][sm] smp support
Update libsm for smp-aware lk kernel.

Replaces the ns-switch thread with two ns-switch threads per cpu and
a single stdcall thread. Each cpu gets an idle-ns-switch thread that
returns to non-secure when no other threads are runable on that cpu
(like the old ns-switch thread) and an irq-ns-switch thread.

The irq-ns-switch thread runs at the highest priority (and with
interrupts masked) and is woken up when an interrupt for the
non-secure os triggers. This thread is used, instead of switching in
the interrupt handler, so that the thread that was running when the
interrupt triggered does not get trapped on the current cpu. The
non-secure os may switch to a higher priority (non-secure-os-)thread
on the current cpu, and call back into the secure os on another cpu.

Both types of switcher threads handle fastcalls directly and wake up
the stdcall thread to handle stdcalls. Only a single stdcall is
allowed at a time, so the return value from the std-call can be
returned to the non-secure os on a different cpu than the cpu that
started the std-call. A new nop stdcall is added that bypass this
restriction.

Also allow secondary cpus to be started after returning to ns on the
main cpu. This does not work well if an interrupt for the non-secure
os triggers before the init hooks are finished.

Change-Id: I637b205fa569bda9dbd11c994d142401ea4ed86c