[lib][sm] Add separate fiq return and restart operations.
Allows returning to the non-secure os to handle FIQs without also
handling normal interrupts. Avoids reporting inaccurate stack traces.

Also avoids reentering trusty on a different cpu after the fiq return.
This is needed for upcoming smp changes so a trusy thread does not get
stuck on a cpu that is running high priority thread in the non-secure
os.

Change-Id: Ifbf7a53ca162ffd925fb21fd6c4033fe51c2dbc6