[lib][trusty][ipc] Fix a bug in handle_list_add
This CL fixes the case when caller adds new handle to the
handle list that some other thread is already waiting on.
In this case it is not sufficient  just to add handle to the list,
we have to also call wait prepare on this new handle to attach event,
check if this new handle is already signaled and wake up waiter thread
if it is.

Change-Id: I43c04c2ebc9e7ae6c37bf50a7a7fb560e2067112