[lib][trusty] Cleanup existing syscall definition
Do not use user_addr_t in user space, just use native types.
Shift * to arg name to be consistent for source.
Do not use uint and unsigned long in favor of fixed width types
when defining syscalls.
Move mm.h to uapi/mm.h - it is a better place for that.

Change-Id: I57ef661b705504c392420d8cbd44c8a02f9b3328