[lib][trusty][handle] Change handle poll prototype
Instead of implemnting finalize_event callback machanism,
which might create a race condition in certain cases, pass
an additional paramater to poll callback to indicate
that it is called in context that guaranties that resulting
event will be delivered to waiter thread so an implementation
can update it's internal state to reflect that if required.

Add event mask parameter for poll call. Only events that
allowed by this mask should be returned to caller.

Change-Id: I092ae1dac8be69a373cbb1575a85d5ec53c7486f