[lib][sm] Use r3 in SWITCH_SCR_TO_NONSECURE before return to ns
Since r4 will be used as variable in NS world by some compiler. 
In this situation the r4 will be changed by SWITCH_SCR_TO_NONSECURE 
to be 0x05 before return to NS world. So that it will pollute the 
NS variabe environment.

Use r3 in this process of SWITCH_SCR_TO_NONSECURE will fix the bug.

Change-Id: Ic95c0a5c365c792da7593bc28f450c2040447ac5
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>