[lib][sm] Use r3 in SWITCH_SCR_TO_NONSECURE before return to ns
Since r4 will be used as variable in NS world by some compiler.
In this situation the r4 will be changed by SWITCH_SCR_TO_NONSECURE
to be 0x05 before return to NS world. So that it will pollute the
NS variabe environment.

Use r3 in this process of SWITCH_SCR_TO_NONSECURE will fix the bug.

Change-Id: I6de74945efc1d4893cf404b87bf413a36ef53f6a
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>