[lib][trusty][app] Move app resource allocation to app start time
App resources (address space, thread, etc) are now allocated right
before the app is started instead of when the app image is parsed.
This allows us to reduce the memory pressure on the system (i.e.
allocating only what is needed at a given time vs. allocating
everything that could be needed).

Change-Id: I376d1a2f17078b3c900ac04418d61e71cb7cee8e