[lib][trusty][app] Add support for start-on-connection and deferred start
A new list of ports is created when parsing application's manifests. This list
is populated with all the the ports specified in applications' manifest with the
START_PORT key. When a connection is made to a port that is not active but that
matches on of the ports in this list, the application that registered the port
is started. Additionally, applications will be restarted after exiting if there
exists a pending connection to any of the ports declared in their manifest.

Applications that do not want to be started at boot (pressumably to be started
later with a connection) can use the TRUSTY_APP_MGMT_FLAGS_DEFERRED_START flag.

Change-Id: Ic012cfaa495f7a3a19c530366826563ecba3685a